Add instructions to people in place pop up, move instructions widget to standalone file.
